Add FreeCell solver, hints, and auto-solve to Blazor UI
- Integrate a full-featured FreeCell solver for move hinting and auto-solving.
- Add "Show Hint" and "Auto-Solve" UI controls with adjustable delay.
- Display solver move suggestions and allow auto-solve cancellation.
- Persist solver settings (auto-move, delay) across sessions.
- Add keyboard shortcuts for solver features (<kbd>H</kbd>, <kbd>S</kbd>).
- Move solver code to production namespace; no breaking changes to gameplay.
- Improve UI/UX with banners, styling, and accessibility for solver features.
