--- TestProject1/FreeCellMove.cs
@@ -1,155 +1,0 @@
-using Client.Games.Cards.Services;
-using Client.Games.Cards.Models;
-
-namespace TestProject1
-{
-    public class FreeCellMove
-    {
-        /*
-         * A move can be between any combination of Foundation/Freecell/Tableau
-         * If tableau to tableau, source and target index are column indexes and cardCount is how many cards from the bottom of the source column
-         * The Seq count for a board is the sum of all sequences that are valid (consecutive, altering red/black). Increasing overall seq count is good
-         * Some moves may move a card to the foundation or freecell, which reduces seq count
-         */
-        public SourceType sourceType { get; set; }
-        public SourceType targetType { get; set; }
-        public int sourceIndex { get; set; } // column # or index of FreeCell or Foundation
-        public int targetIndex { get; set; } // column # or index of FreeCell or Foundation
-        public int srcColumnIndex { get; set; } // only for tableau to tableau moves
-        public int cardCount { get; set; } // only for tableau to tableau moves, how many cards from the bottom of the source column
-
-        public int Depth { get; set; } // from 0. Depth of the move in the search tree, used for debugging and logging
-        public bool DidExecuteMove { get; set; } = false;
-        public FreeCellMove? ParentMove { get; set; } // used for debugging and logging to trace back the sequence of moves that led to this move
-        public List<FreeCellMove> ChildMoves { get; set; } = new List<FreeCellMove>(); // used for debugging and logging to trace the sequence of moves that led to this move and the moves that can be made from this move
-
-        public int mValue { get; set; } // move value
-        public int deltaBValue { get; set; }
-        public Card? CardMoved { get; set; }
-        // Sequence-clear heuristic: remaining moves in a multi-card sequence are queued here
-        // so the backtracker can explore alternatives at each step rather than being locked in.
-        public Queue<FreeCellMove>? PendingSequenceMoves { get; set; }
-        public bool IsRootNode => ParentMove == null;
-        public FreeCellMove(Card? cardMoved)
-        {
-            CardMoved = cardMoved;
-        }
-        /// <summary>
-        /// Creates a new FreeCellMove with source and target swapped, representing the reverse of this move.
-        /// </summary>
-        public FreeCellMove CreateReversedMove() => new FreeCellMove(CardMoved)
-        {
-            sourceType = targetType,
-            targetType = sourceType,
-            sourceIndex = targetIndex,
-            targetIndex = sourceIndex,
-            cardCount = cardCount
-        };
-
-        public bool ApplyMove(FreeCellGameBase game)
-        {
-            var cardIndex = -1;
-            if (sourceType == SourceType.Tableau)
-            {
-                cardIndex = game.Tableau[sourceIndex].Count - cardCount;
-            }
-            // use the TryMove method to apply this move to the game in memory
-            game.Selection = new CardSelection
-            {
-                SourceType = sourceType,
-                SourceIndex = sourceIndex,
-                CardIndex = cardIndex // not used for tableau to tableau moves since we always move from the bottom of the column, and not used for freecell or foundation moves since they only have one card
-            };
-            var didMove = game.TryMove(targetType, targetIndex);
-            return didMove;
-        }
-
-        /// <summary>
-        /// Optimization: fast-path move application that bypasses TryMove validation.
-        /// Directly manipulates game state — only safe when the move is known-valid
-        /// (e.g., generated by the solver's move finder which already validated legality).
-        /// </summary>
-        public bool ApplyMoveFast(FreeCellGameBase game)
-        {
-            bool inc = game.IncrementalHashReady;
-            switch (sourceType)
-            {
-                case SourceType.FreeCell:
-                    var fc = game.FreeCells[sourceIndex];
-                    if (fc == null) return false;
-                    if (inc) game.HashRemoveFromFreeCell(sourceIndex);
-                    game.FreeCells[sourceIndex] = null;
-                    PlaceCardsFast(game, [fc], inc);
-                    break;
-
-                case SourceType.Tableau:
-                    var srcCol = game.Tableau[sourceIndex];
-                    var startIdx = srcCol.Count - cardCount;
-                    if (inc) game.HashRemoveFromTableau(sourceIndex, startIdx, cardCount);
-                    if (cardCount == 1)
-                    {
-                        var card = srcCol[startIdx];
-                        srcCol.RemoveAt(startIdx);
-                        PlaceCardsFast(game, [card], inc);
-                    }
-                    else
-                    {
-                        var cards = srcCol.GetRange(startIdx, cardCount);
-                        srcCol.RemoveRange(startIdx, cardCount);
-                        PlaceCardsFast(game, cards, inc);
-                    }
-                    break;
-
-                case SourceType.Foundation:
-                    var fPile = game.Foundations[sourceIndex];
-                    var fCard = fPile[^1];
-                    if (inc) game.HashRemoveFromFoundation(sourceIndex);
-                    fPile.RemoveAt(fPile.Count - 1);
-                    PlaceCardsFast(game, [fCard], inc);
-                    break;
-
-                default:
-                    return false;
-            }
-            game.MoveCount++;
-            return true;
-        }
-
-        private void PlaceCardsFast(FreeCellGameBase game, List<Card> cards, bool inc)
-        {
-            switch (targetType)
-            {
-                case SourceType.FreeCell:
-                    game.FreeCells[targetIndex] = cards[0];
-                    if (inc) game.HashAddToFreeCell(targetIndex);
-                    break;
-                case SourceType.Tableau:
-                    int oldCount = game.Tableau[targetIndex].Count;
-                    game.Tableau[targetIndex].AddRange(cards);
-                    if (inc) game.HashAddToTableau(targetIndex, oldCount, cards.Count);
-                    break;
-                case SourceType.Foundation:
-                    game.Foundations[targetIndex].Add(cards[0]);
-                    if (inc) game.HashAddToFoundation(targetIndex);
-                    break;
-            }
-        }
-        public override string ToString() => $"{CardMoved} {sourceType}[{sourceIndex}]->{targetType}[{targetIndex}] cards:{cardCount} mVal:{mValue} {(DidExecuteMove ? "!" : "")}";
-
-        /// <summary>
-        /// Reverses this move by creating the reversed move and applying it.
-        /// Uses ApplyMoveFast on the reversed move, then adjusts MoveCount to net -1.
-        /// </summary>
-        public bool UnApplyMove(FreeCellGameBase game)
-        {
-            var reversed = CreateReversedMove();
-            if (!reversed.ApplyMoveFast(game))
-            {
-                return false;
-            }
-            // ApplyMoveFast increments MoveCount by 1, but unapply should decrement by 1 → net adjustment of -2
-            game.MoveCount -= 2;
-            return true;
-        }
-    }
-}
